+++ macros/src/actix.rs
@@ -649,7 +649,7 @@ pub fn emit_v2_security(input: TokenStream) -> TokenStream {
         security_attrs.get("parent"),
     ) {
         (Some(type_), None) => {
-            let alias = security_attrs.get("alias").unwrap_or_else(|| type_);
+            let alias = security_attrs.get("alias").unwrap_or(type_);
             let quoted_description = quote_option(security_attrs.get("description"));
             let quoted_name = quote_option(security_attrs.get("name"));
             let quoted_in = quote_option(security_attrs.get("in"));