+++ plugins/actix-web/src/lib.rs
@@ -1,3 +1,8 @@
+#[cfg(feature = "actix2")]
+extern crate actix_web2 as actix_web;
+#[cfg(feature = "actix3")]
+extern crate actix_web3 as actix_web;
+
 pub mod web;
 
 pub use self::web::{Resource, Route, Scope};