+++ src/lib.rs
@@ -17,7 +17,7 @@ pub use paperclip_core::util;
 #[cfg(feature = "v2")]
 pub use paperclip_macros::api_v2_schema_struct as api_v2_schema;
 
-#[cfg(feature = "actix")]
+#[cfg(feature = "actix-base")]
 pub mod actix {
     //! Plugin types, traits and macros for actix-web framework.